+++ CHANGELOG.md
@@ -129,7 +129,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [1.0.0] - 2026-05-13
 
-This marks the first stable, production-ready release of the Scotch Egg Bot!
+This marks the first stable, production-ready release of Scotch Egg!
 
 ### Added